Add option to hide overview button
For some entries with multiple storylines or very short entries, the
overview in its current form does not make sense. Add an option to
hide the overview button in navigation bars.

Whether the option is available at all, can be controlled via a theme
option `no_overview_button`. This replaces the SCSS based hiding of
the overview button.