Make memberships' entity polymorphic, require role
Memberships can now refer to accounts as well as entrys via the
polymorphic entity attribute, which replaces the entry attribute.

Memberships now also have a non-null role attribute which defaults to
'editor'. This attribute will also hold the values 'admin', 'publisher'
and 'reader' as soon as the permissions model is replaced.

Memberships as of this commit cannot be created or deleted in the admin
interface via their user. Ways to facilitate a desired interaction
between Formtastic, ActiveAdmin and the polymorphic entity attribute
might be found at
https://github.com/activeadmin/activeadmin/wiki/Nested-forms-with-polymorphic-association-in-Active-Admin-Formtastic
.