Remove hints about invisible users or accounts
While it is possible that a user sees memberships of users that do not
show up in his users admin, always displaying a hint leads to
confusion for users that do not face this situation.

Additionally, the hint was always displayed soon as there was an
account that the user is not manager of. Still, for users that are
managers of all there accounts, everything looks absolutely
consistent.