Make CollectionView sorting more robust
While adding an item to the collection, the sort event can fire
multiple times and even before the add event has been handled by the
collection view. In that case we have to be able to reorder the item
views even if there are collection items for which no item views exist
yet.