Improve hls options fallback
Also fall back to `s3_host_alias` when `hls_host_alias` or
`hls_origin_host_alias` is `nil`, not only when the key is not
present.

Otherwise, when reading the option from an environment variable, we
have to make sure not to set the key if the value is not present.