Ensure connections for removed video elements are closed
When removing a video element, Firefox 67 keeps an open connection for
the src. It can take over a minute for Firefox to finish the request.
When viewing multiple videos in a row, the count of open connections
can hit `Network.http.max-persistent-connections-per-server` causing
subsequent videos not to load.

Resetting video src to a blank file appears to cause the connection to
be closed.

REDMINE-16834